+++ app/src/main/java/it/unimib/brewday/util/Costanti.java
@@ -2,5 +2,6 @@ package it.unimib.brewday.util;
 
 public class Costanti {
 
+    private Costanti(){}
     public static final String DATABASE_NAME = "brew-day-db";
 }
